+++ origin/tests/TestCase/Core/SessionTest.php
@@ -18,63 +18,60 @@ use Origin\Core\Session;
 
 class SessionTest extends \PHPUnit\Framework\TestCase
 {
+    public function setUp()
+    {
+        $this->Session = new Session();
+    }
+ 
     public function testWrite()
     {
-        Session::write('sessionTest', 'works');
+        $this->Session->write('sessionTest', 'works');
         $this->assertTrue(isset($_SESSION['sessionTest']));
         $this->assertEquals('works', $_SESSION['sessionTest']);
     }
 
     public function testRead()
     {
-        $this->assertNull(Session::read('sessionTest'));
+        $this->assertNull($this->Session->read('sessionTest'));
+
+        $this->Session->write('sessionTest', 'works');
+
+        $this->assertEquals('works', $this->Session->read('sessionTest'));
         
-        Session::write('sessionTest', 'works');
-        $this->assertEquals('works', Session::read('sessionTest'));
-        
-        Session::write('Test.status', 'ok');
-        $this->assertEquals('ok', Session::read('Test.status'));
+        $this->Session->write('Test.status', 'ok');
+        $this->assertEquals('ok', $this->Session->read('Test.status'));
     }
 
     public function testCheck()
     {
-        Session::write('Test.status', 'ok');
-        $this->assertTrue(Session::check('Test.status'));
-        $this->assertFalse(Session::check('Test.password'));
+        $this->Session->write('Test.status', 'ok');
+        $this->assertTrue($this->Session->check('Test.status'));
+        $this->assertFalse($this->Session->check('Test.password'));
     }
 
     public function testDelete()
     {
-        Session::write('Test.status', 'ok');
-        $this->assertTrue(Session::delete('Test.status'));
-        $this->assertFalse(Session::delete('Test.password'));
+        $this->Session->write('Test.status', 'ok');
+        $this->assertTrue($this->Session->delete('Test.status'));
+        $this->assertFalse($this->Session->delete('Test.password'));
     }
 
     public function testDestroy()
     {
-        Session::write('Test.status', 'ok');
+        $this->Session->write('Test.status', 'ok');
 
-        $this->assertTrue(Session::started());
-        Session::destroy();
-        $this->assertFalse(Session::check('Test.status'));
+        $this->assertTrue($this->Session->started());
+        $this->Session->destroy();
+        $this->assertFalse($this->Session->check('Test.status'));
     }
     /**
      * @depends testDestroy
      */
     public function testCreate()
     {
-        Session::destroy();
-        Session::initialize();
-        Session::write('Test.status', 'ok');
-        $this->assertTrue(Session::check('Test.status'));
-        $this->assertTrue(Session::check('Session.lastActivity'));
-    }
-
-    public function testTimeout()
-    {
-        Session::write('Test.status', 'ok');
-        Session::write('Session.lastActivity', 0);
-        Session::initialize();
-        $this->assertFalse(Session::check('Test.status'));
+        $this->Session->destroy();
+        $this->Session->start();
+        $this->Session->write('Test.status', 'ok');
+        $this->assertTrue($this->Session->check('Test.status'));
     }
 }